requirement.txt
streamlit
tensorflow==2.18.0
numpy
pillow
keras